% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_extract.R
\name{logger_extract}
\alias{logger_extract}
\title{Extract AIMS MMP Logger Data}
\usage{
logger_extract(
  SelectedYears,
  SelectedLoggers,
  FilterFlags = TRUE,
  FlagTags = c(1, 2),
  AggregateToDaily = FALSE
)
}
\arguments{
\item{SelectedYears}{Numeric Vector. The year(s) of data you want to access.}

\item{SelectedLoggers}{Character Vector. The name(s) of the logger(s) you want to access.}

\item{FilterFlags}{Boolean. Do you want to filter the data by its quality flags? This defaults to TRUE (yes).}

\item{FlagTags}{Numeric Vector. A vector of flag tags to keep, choices are: 0 =  No_QC_performed,
1 = Good_data, 2 = Probably_good_data, 3 = Bad_data_that_are_potentially_correctable, 4 = Bad_data, 5 = Value_changed.
Advice from data providers is to keep only tags 1 and 2. By default only tags 1 and 2 are kept.}

\item{AggregateToDaily}{Boolean. Do you want to aggregate data to daily values? This defaults to FALSE (no) and returns
15-minute interval data.}
}
\value{
A long format dataframe
}
\description{
Extract AIMS MMP Logger Data
}
\examples{
wq_data <- logger_extract(
SelectedYears = 2025, 
SelectedLoggers = "BUR2", 
FilterFlags = TRUE, 
FlagTags = c(1,2), 
AggregateToDaily = FALSE)
}
